sys - Move Fcb functions under util folder
