+++ sys/close.c
@@ -21,6 +21,7 @@ with this program. If not, see <http://www.gnu.org/licenses/>.
 */
 
 #include "dokan.h"
+#include "util/fcb.h"
 
 NTSTATUS
 DokanDispatchClose(__in PDEVICE_OBJECT DeviceObject, __in PIRP Irp)
