+++ sys/sys.vcxproj.filters
@@ -100,6 +100,9 @@
     <ClCompile Include="util\log.c">
       <Filter>Source Files\util</Filter>
     </ClCompile>
+    <ClCompile Include="util\fcb.c">
+      <Filter>Source Files\util</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="dokan.h">
@@ -115,6 +118,9 @@
       <Filter>Header Files\util</Filter>
     </ClInclude>
     <ClInclude Include="util\log.h">
+      <Filter>Header Files\util</Filter>
+    </ClInclude>
+    <ClInclude Include="util\fcb.h">
       <Filter>Header Files\util</Filter>
     </ClInclude>
   </ItemGroup>
